refactor(cli): extract shared client factory and settings hierarchy
- Add ICliKustoClientFactory/CliKustoClientFactory for shared Kusto client creation
- Introduce ClusterCommandSettings and DatabaseCommandSettings base classes
- Reparent ExportBaseCommandSettings to use new settings hierarchy
- Refactor KustoSchemaExporter to use ICliKustoClientFactory instead of managing client internally
- Register ICliKustoClientFactory as singleton in Program.cs
