+++ src/Atc.Kusto.CLI/Atc.Kusto.CLI.csproj
