+++ src/Atc.Kusto.CLI/appsettings.json
